% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_models.R
\name{adjust_models}
\alias{adjust_models}
\title{Adjust K linear mixed effects models}
\usage{
adjust_models(data, formula, K, clusters)
}
\arguments{
\item{data}{data frame containing the variables named in \code{formula}.}

\item{formula}{a two sided linear formula object corresponding to \code{formula} parameter
in \code{lmer} from \pkg{lme4} package.}

\item{K}{number of clusters.}

\item{clusters}{vector corresponding to the clustering.}
}
\value{
a class "clmm" object containing the list of the adjusted K linear mixed effects models.
}
\description{
Define and adjust a linear mixed effects model for each cluster given a clustering.
Fitting the models is done using \pkg{lme4} package with default parameters.
}
