% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{MSE error}
\usage{
mse(prediction, target, residuals = FALSE)
}
\arguments{
\item{prediction}{vector containing the response of the clusterwise lmm to the data.}

\item{target}{vector containing the target variable.}

\item{residuals}{logical; if TRUE, returns the squared of residuals instead of the MSE}
}
\value{
mean squared error between prediction and target
}
\description{
Compute cost function for clusterwise lmm that is the mean squared error between
the prediction and the target variable
}
