% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_batch.R
\name{optimizer_batch}
\alias{optimizer_batch}
\title{Fitting clusterwise linear mixed effects model on a given data}
\usage{
optimizer_batch(
  data,
  target,
  K,
  formula,
  nb_iterations = 100,
  initialization = NULL
)
}
\arguments{
\item{data}{data frame containing the variables named in \code{formula}.}

\item{target}{vector containing the target variable.}

\item{K}{number of clusters.}

\item{formula}{a two sided linear formula object corresponding to \code{formula} parameter
in \code{lmer} from \pkg{lme4} package.}

\item{nb_iterations}{maximum number of iterations.}

\item{initialization}{if vector of the same length of data, use this vector as clustering initialization.}
}
\value{
list containing the models as a class of clmm, the final clustering and the final cost.
}
\description{
run clusterwise linear mixed effects model with K clusters on the given data. This
function uses a batch approach where the cost are calculated along all rows and the affectation is
made to minimize this cost, after that the k models are adjusted.
}
