% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cluster.R
\name{predict_cluster}
\alias{predict_cluster}
\title{Predict an optimal clustering}
\usage{
predict_cluster(data, target, models)
}
\arguments{
\item{data}{data frame containing the variables in which the clmm model has been trained.}

\item{target}{vector containing the target variable.}

\item{models}{object of class clmm that contains a list of k lmm models.}
}
\value{
vector containing the optimal clustering.
}
\description{
Compute and return an optimal clustering that minimize the MSE error given K lmm models.
}
