% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate 3 levels data}
\usage{
simulate_data(
  n1,
  n2,
  n3,
  fixed_slope,
  fixed_intercept,
  lvl2_corr_mat,
  lvl3_corr_mat,
  sigma2,
  separable_in_x = TRUE,
  distance_factor = 2
)
}
\arguments{
\item{n1}{number of individuals at level 1. Must be strictly positive.}

\item{n2}{number of subgroups at level 2. Must be strictly positive.}

\item{n3}{number of groups at level 3. Must be strictly positive.}

\item{fixed_slope}{Fixed effect slope. Must be finite.}

\item{fixed_intercept}{Fixed effect intercept. Must be finite.}

\item{lvl2_corr_mat}{Level 2 random effect correlation matrix. Must be of type
\pkg{matrix} with size of 2x2}

\item{lvl3_corr_mat}{Level 3 random effect correlation matrix. Must be of type
\pkg{matrix} with size of 2x2}

\item{sigma2}{variance of the residuals. Must be strictly positive.}

\item{separable_in_x}{logical; if TRUE, the data will be separable in x axis.}

\item{distance_factor}{numeric; When separable_in_x == TRUE, indicates the distances in
x between level 3 groups. Must be strictly positive.}
}
\value{
data frame containing the generated data containing the following list of columns :\cr
\itemize{
\item x : explicative variable\cr
\item y : target variable\cr
\item lvl2 : factors indicating level 2 groups\cr
\item lvl3 : factors indicating level 3 groups\cr
\item lvl3_intercepts : numeric indicating level 3 random intercepts\cr
\item lvl3_slopes : numeric indicating level 3 random slopes\cr
\item lvl2_intercepts : numeric indicating level 2 random intercepts\cr
\item lvl2_slopes : numeric indicating level 2 random slopes\cr
}
}
\description{
Generate univariate (for now) linear 3 level data with \code{n3} groups,
\code{n2} subgroups per group and \code{n1} individuals per subgroups.
}
\details{
The generated data follows the 3 levels lmm equations given below :
\deqn{Y_{ijk} = \beta_{0jk} + \beta_{1jk} X_{ijk} + \epsilon_{ijk}}
Level 2 :
\deqn{\beta_{0jk} = \gamma_{00k} + u_{0jk}}
\deqn{\beta_{1jk} = \gamma_{10k} + u_{1jk}}
Level 3 :
\deqn{\gamma_{00k} = \delta_{000} + u_{00k}}
\deqn{\gamma_{10k} = \delta_{100} + u_{10k}}
With : \cr
\itemize{
\item \eqn{\delta_{000}} and \eqn{\delta_{100}} : fixed effects intercept and slope.\cr
\item \eqn{u_{00k}} and \eqn{u_{10k}} : level 3 random effects intercept and slope for each level 3 group.\cr
\item \eqn{u_{0jk}} and \eqn{u_{0jk}} : level 2 random effects intercept and slope for each level 2 group within level 3 group.\cr
\item \eqn{\epsilon_{ijk}} : residual noise.
}
}
\examples{
## define correlation matrices for random effects slopes and intercepts
lvl2_corr_mat = matrix(c(1, 0,
                         0, 1), 2, 2)
lvl3_corr_mat = matrix(c(10, 0,
                         0, 10), 2, 2)

## to generate 3 levels data with 3 groups that contains each 10 subgroups,
## and each subgroup contains 30 individuals
data = simulate_data(n1 = 30,
                     n2 = 10,
                     n3 = 3,
                     fixed_slope = -2,
                     fixed_intercept = 2,
                     lvl2_corr_mat = lvl2_corr_mat,
                     lvl3_corr_mat = lvl3_corr_mat,
                     sigma2 = 1
                     )

}
