% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_example.R
\name{testing_example}
\alias{testing_example}
\title{Testing clusterwise lmm}
\usage{
testing_example()
}
\value{
shows plots for the level 3 clustering and the resulting one
}
\description{
This function runs \code{clusterwiselmm} function on a simulated data set
and illustrate the resulting clustering. you will find this example in the vignette.
You can also change the parameters of the simulation.
}
\examples{
testing_example()
}
